WIP tool to generate sprites from game data
Currently can extract sprites from PNG and dump them to a directory
The problem is some of the game's files are missing; notably PNGs.
The solution is there are ATF files from which sprites can be extracted.
The problem is there are no great off-the-shelf ATF2PNG tools; they all
don't work or only work for specific types of files.
I'm working on the solution there (see Models/AtfFile.cs...)