Fix the remaining decoding bugs regarding alpha
The major issue was alpha index selection was wrong; that's now fixed by
making the alpha get applied to pixelIndex `3 - y`. Changing it to `3 -
y` broke the color selection but making the color index extraction
little endian fixed that up.
The other issue that was very straight forward was that a few commits
ago I removed the hardcoded BuildAlphas() in favor of a loop mechanism
but I forgot I had tested reversing the lerping direction (which was
wrong); this commit reverts that back to the proper lerp order